+++ packages/interactive_media_ads/CHANGELOG.md
@@ -1,3 +1,15 @@
+## 0.2.6+2
+
+* Updates kotlin version to 2.2.0 to enable gradle 8.11 support.
+
+## 0.2.6+1
+
+* Fixes passing ads response to Android native `AdsRequest`.
+
+## 0.2.6
+
+* Adds support to configure ad requests. See `AdsRequest`.
+
 ## 0.2.5+1
 
 * Adds remaining methods for internal wrapper of the Android native `AdsRequest`.