+++ packages/webview_flutter/webview_flutter_android/CHANGELOG.md
@@ -1,6 +1,15 @@
-## 4.9.0
+## 4.10.0
 
 * Adds support for the Payment Request API with `AndroidWebViewController.isWebViewFeatureSupported` and `AndroidWebViewController.setPaymentRequestEnabled`.
+
+## 4.9.1
+
+* Updates kotlin version to 2.2.0 to enable gradle 8.11 support.
+
+## 4.9.0
+
+* Adds support for `PlatformWebViewController.loadFileWithParams`.
+* Introduces `AndroidLoadFileParams`, a platform-specific extension of `LoadFileParams` for Android that adds support for `headers`.
 
 ## 4.8.2